airflow-docker/requirements.txt
tqdm==4.66.4
aiosqlite==0.20.0
psycopg2-binary
SQLAlchemy>=1.4.0
pandas
requests>=2.31.0
apache-airflow-providers-postgres==5.7.1
GeoAlchemy2>=0.13.0
stem==1.8.1
fake-useragent==1.2.1
aiohttp-socks==0.8.4
aiohttp==3.9.1
async-timeout==4.0.3
beautifulsoup4==4.12.2
python-dotenv==1.0.0
python-socks[asyncio]>=2.0.0